test(CheckMonitoredTrip): Move one-time trip test from CheckMonitoredTripBasic
